scoring_service/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client and web scraping
httpx==0.25.2
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# PDF processing and OCR
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0

# Text processing and similarity
rapidfuzz==3.5.2
sentence-transformers==2.2.2
nltk==3.8.1
unidecode==1.3.7

# Data processing
pandas==2.1.4
numpy==1.25.2

# Caching and queues
redis==5.0.1
celery==5.3.4

# Storage
boto3==1.34.0
minio==7.2.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Security and auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0
